% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_fill_module.R
\name{dateconverterRdfpJst}
\alias{dateconverterRdfpJst}
\title{dateconverterRdfpJst}
\usage{
dateconverterRdfpJst(this_date)
}
\arguments{
\item{this_date}{the date to be converted}
}
\value{
returns a list representing the date
}
\description{
converts dates to a list that the DFP API understnads
}
\examples{
dateconverterRdfpJst(as.Date("2019-01-01"))
}
