% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_fill_module.R
\name{df_ghfill_mock_creation}
\alias{df_ghfill_mock_creation}
\title{df_ghfill_mock_creation}
\usage{
df_ghfill_mock_creation()
}
\description{
creates up a dummy tibble with 0 values
}
\examples{
df_ghfill_mock_creation()
}
\keyword{none}
