% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_fill_module.R
\name{gh_fill}
\alias{gh_fill}
\title{gh_fill}
\usage{
gh_fill(input, output, session)
}
\arguments{
\item{input}{will be filled automatically by shiny}

\item{output}{will be filled automatically by shiny}

\item{session}{will be filled automatically by shiny}
}
\value{
returns code to be used in shiny server file
}
\description{
Server part of ghanafill shinyModule
}
