% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_fill_module.R
\name{gh_fillUI}
\alias{gh_fillUI}
\title{gh_fillUI}
\usage{
gh_fillUI(id)
}
\arguments{
\item{id}{use a unique id for the module}
}
\value{
returns a taglist to be used in shiny UI
}
\description{
UI part of ghanafill shinyModule
}
